Move cell operations to its own module
